Add sanitize compilation rule and initialize 1st node of every list
The uninitialization lead to "SEGV on unknown address [...] caused by a READ memory access [...] in ft_lstadd_back" because it was accessing a not NULL but uninitialized first element of the list in the condition of the first if in the mentioned function.

This segmentation fault was hidden and apparently not causing any harm, but detected using fsanitizer flag.